refactor: PostListSkeleton features내부로 이동
